+++ pytest_selenium/pytest_selenium.py
@@ -11,17 +11,20 @@ import logging
 import json
 
 import pytest
-from requests.structures import CaseInsensitiveDict
 from selenium import webdriver
 from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
 from selenium.webdriver.support.event_firing_webdriver import EventFiringWebDriver
+from tenacity import Retrying, stop_after_attempt, wait_exponential
 
+from .utils import CaseInsensitiveDict
 from . import drivers
 
 import allure
 
 LOGGER = logging.getLogger(__name__)
+import warnings
 
+LOGGER = logging.getLogger(__name__)
 
 SUPPORTED_DRIVERS = CaseInsensitiveDict(
     {
@@ -31,7 +34,6 @@ SUPPORTED_DRIVERS = CaseInsensitiveDict(
         "Edge": webdriver.Edge,
         "Firefox": webdriver.Firefox,
         "IE": webdriver.Ie,
-        "PhantomJS": webdriver.PhantomJS,
         "Remote": webdriver.Remote,
         "Safari": webdriver.Safari,
         "SauceLabs": webdriver.Remote,
@@ -93,7 +95,12 @@ def session_capabilities(pytestconfig):
 
 @pytest.fixture
 def capabilities(
-    request, driver_class, chrome_options, firefox_options, session_capabilities
+    request,
+    driver_class,
+    chrome_options,
+    firefox_options,
+    edge_options,
+    session_capabilities,
 ):
     """Returns combined capabilities"""
     capabilities = copy.deepcopy(session_capabilities)  # make a copy
@@ -108,6 +115,9 @@ def capabilities(
         elif browser == "FIREFOX":
             key = firefox_options.KEY
             options = firefox_options.to_capabilities()
+        elif browser == "EDGE":
+            key = edge_options.KEY
+            options = edge_options.to_capabilities()
         if all([key, options]):
             capabilities[key] = _merge(capabilities.get(key, {}), options.get(key, {}))
     capabilities.update(get_capabilities_from_markers(request.node))
@@ -143,6 +153,7 @@ def driver_kwargs(
     driver_path,
     firefox_options,
     firefox_profile,
+    edge_options,
     pytestconfig,
 ):
     kwargs = {}
@@ -156,13 +167,15 @@ def driver_kwargs(
             driver_path=driver_path,
             firefox_options=firefox_options,
             firefox_profile=firefox_profile,
-            host=pytestconfig.getoption("host"),
-            port=pytestconfig.getoption("port"),
+            edge_options=edge_options,
+            host=pytestconfig.getoption("selenium_host"),
+            port=pytestconfig.getoption("selenium_port"),
             service_log_path=None,
             request=request,
             test=".".join(split_class_and_test_names(request.node.nodeid)),
         )
     )
+
     pytestconfig._driver_log = driver_log
     return kwargs
 
@@ -189,7 +202,16 @@ def driver_path(request):
 @pytest.fixture
 def driver(request, driver_class, driver_kwargs):
     """Returns a WebDriver instance based on options and capabilities"""
-    driver = driver_class(**driver_kwargs)
+
+    retries = int(request.config.getini("max_driver_init_attempts"))
+    for retry in Retrying(
+        stop=stop_after_attempt(retries), wait=wait_exponential(), reraise=True
+    ):
+        with retry:
+            LOGGER.info(
+                f"Driver init, attempt {retry.retry_state.attempt_number}/{retries}"
+            )
+            driver = driver_class(**driver_kwargs)
 
     event_listener = request.config.getoption("event_listener")
     if event_listener is not None:
@@ -212,6 +234,22 @@ def selenium(driver):
 
 @pytest.hookimpl(trylast=True)
 def pytest_configure(config):
+    if config.getoption("host"):
+        warnings.warn(
+            "--host has been deprecated and will be removed in a "
+            "future release. Please use --selenium-host instead.",
+            DeprecationWarning,
+        )
+        config.option.selenium_host = config.getoption("host")
+
+    if config.getoption("port"):
+        warnings.warn(
+            "--port has been deprecated and will be removed in a "
+            "future release. Please use --selenium-port instead.",
+            DeprecationWarning,
+        )
+        config.option.selenium_port = config.getoption("port")
+
     capabilities = config._variables.get("capabilities", {})
     capabilities.update({k: v for k, v in config.getoption("capabilities")})
     config.addinivalue_line(
@@ -225,9 +263,9 @@ def pytest_configure(config):
     if hasattr(config, "_metadata"):
         config._metadata["Driver"] = config.getoption("driver")
         config._metadata["Capabilities"] = capabilities
-        if all((config.getoption("host"), config.getoption("port"))):
+        if all((config.option.selenium_host, config.option.selenium_port)):
             config._metadata["Server"] = "{0}:{1}".format(
-                config.getoption("host"), config.getoption("port")
+                config.option.selenium_host, config.option.selenium_port
             )
     config._capabilities = capabilities
 
@@ -367,7 +405,11 @@ def _gather_logs(item, report, driver, summary, extra):
 
 def _gather_driver_log(item, summary, extra):
     pytest_html = item.config.pluginmanager.getplugin("html")
-    if hasattr(item.config, "_driver_log") and os.path.exists(item.config._driver_log):
+    if (
+        hasattr(item.config, "_driver_log")
+        and item.config._driver_log is not None
+        and os.path.exists(item.config._driver_log)
+    ):
         if pytest_html is not None:
             with io.open(item.config._driver_log, "r", encoding="utf8") as f:
                 extra.append(pytest_html.extras.text(f.read(), "Driver Log"))
@@ -377,7 +419,7 @@ def _gather_driver_log(item, summary, extra):
 def format_log(log):
     timestamp_format = "%Y-%m-%d %H:%M:%S.%f"
     entries = [
-        u"{0} {1[level]} - {1[message]}".format(
+        "{0} {1[level]} - {1[message]}".format(
             datetime.utcfromtimestamp(entry["timestamp"] / 1000.0).strftime(
                 timestamp_format
             ),
@@ -405,8 +447,12 @@ class DriverAction(argparse.Action):
         setattr(namespace, self.dest, values)
         driver = getattr(drivers, values.lower())
         # set the default host and port if specified in the driver module
-        namespace.host = namespace.host or getattr(driver, "HOST", None)
-        namespace.port = namespace.port or getattr(driver, "PORT", None)
+        namespace.selenium_host = namespace.selenium_host or getattr(
+            driver, "HOST", None
+        )
+        namespace.selenium_port = namespace.selenium_port or getattr(
+            driver, "PORT", None
+        )
 
 
 def pytest_addoption(parser):
@@ -429,6 +475,11 @@ def pytest_addoption(parser):
         default=os.getenv("SAUCELABS_JOB_AUTH", "none"),
     )
 
+    parser.addini(
+        "max_driver_init_attempts",
+        help="Maximum number of driver initialization attempts",
+        default=3,
+    )
     group = parser.getgroup("selenium", "selenium")
     group._addoption(
         "--driver",
@@ -458,12 +509,27 @@ def pytest_addoption(parser):
     group._addoption(
         "--host",
         metavar="str",
-        help="host that the selenium server is listening on, "
+        help="DEPRECATED host that the selenium server is listening on, "
         "which will default to the cloud provider default "
         "or localhost.",
     )
     group._addoption(
         "--port",
+        type=int,
+        metavar="num",
+        help="DEPRECATED port that the selenium server is listening on, "
+        "which will default to the cloud provider default "
+        "or localhost.",
+    )
+    group._addoption(
+        "--selenium-host",
+        metavar="str",
+        help="host that the selenium server is listening on, "
+        "which will default to the cloud provider default "
+        "or localhost.",
+    )
+    group._addoption(
+        "--selenium-port",
         type=int,
         metavar="num",
         help="port that the selenium server is listening on, "