+++ testing/test_driver.py
@@ -4,7 +4,10 @@
 
 from functools import partial
 
+import pytest_selenium
 import pytest
+
+from pytest_selenium.utils import CaseInsensitiveDict
 
 pytestmark = pytest.mark.nondestructive
 
@@ -103,7 +106,11 @@ def test_default_host_port(testdir):
     testdir.quick_qa("--driver", "Remote", file_test, passed=1)
 
 
-def test_arguments_order(testdir):
+@pytest.mark.parametrize(
+    ("host_arg_name", "port_arg_name"),
+    [("--selenium-host", "--selenium-port"), ("--host", "--port")],
+)
+def test_arguments_order(testdir, host_arg_name, port_arg_name):
     host = "notlocalhost"
     port = "4441"
     file_test = testdir.makepyfile(
@@ -117,7 +124,14 @@ def test_arguments_order(testdir):
         )
     )
     testdir.quick_qa(
-        "--driver", "Remote", "--host", host, "--port", port, file_test, passed=1
+        "--driver",
+        "Remote",
+        host_arg_name,
+        host,
+        port_arg_name,
+        port,
+        file_test,
+        passed=1,
     )
 
 
@@ -135,7 +149,14 @@ def test_arguments_order_random(testdir):
         )
     )
     testdir.quick_qa(
-        "--host", host, "--driver", "Remote", "--port", port, file_test, passed=1
+        "--selenium-host",
+        host,
+        "--driver",
+        "Remote",
+        "--selenium-port",
+        port,
+        file_test,
+        passed=1,
     )
 
 
@@ -180,3 +201,58 @@ def test_no_service_log_path(testdir):
     """
     )
     testdir.quick_qa("--driver", "Firefox", file_test, passed=1)
+
+
+def test_driver_retry_pass(testdir, mocker):
+    mocker.patch.dict(
+        pytest_selenium.SUPPORTED_DRIVERS,
+        CaseInsensitiveDict({"Firefox": mocker.MagicMock()}),
+    )
+    mock_retrying = mocker.spy(pytest_selenium.pytest_selenium, "Retrying")
+
+    file_test = testdir.makepyfile(
+        """
+        import pytest
+
+        @pytest.mark.nondestructive
+        def test_pass(driver):
+            assert driver
+    """
+    )
+
+    testdir.quick_qa("--driver", "Firefox", file_test, passed=1)
+    assert mock_retrying.spy_return.statistics["attempt_number"] == 1
+
+
+@pytest.mark.parametrize("max_num_attempts", [None, 2])
+def test_driver_retry_fail(testdir, mocker, max_num_attempts):
+    mocker.patch(
+        "pytest_selenium.webdriver.firefox.webdriver.FirefoxRemoteConnection",
+        side_effect=Exception("Connection Error"),
+    )
+    mocker.patch("pytest_selenium.webdriver.firefox.webdriver.Service")
+    mocker.patch("pytest_selenium.pytest_selenium.wait_exponential", return_value=0)
+    mock_retrying = mocker.spy(pytest_selenium.pytest_selenium, "Retrying")
+
+    default_attempts = 3
+    if max_num_attempts is not None:
+        testdir.makeini(
+            f"""
+            [pytest]
+            max_driver_init_attempts = {max_num_attempts}
+            """
+        )
+
+    file_test = testdir.makepyfile(
+        """
+        import pytest
+
+        @pytest.mark.nondestructive
+        def test_pass(driver):
+            assert True  # catch if this starts to pass
+        """
+    )
+
+    testdir.quick_qa("--driver", "Firefox", file_test, failed=1)
+    expected_attempts = max_num_attempts or default_attempts
+    assert mock_retrying.spy_return.statistics["attempt_number"] == expected_attempts